* add generator inference script and options * update mnt encoder inference script
